+++ XC8Projects/16F18855/BBQ_Controller.X/tempCalc.h
@@ -3,6 +3,6 @@
 
 #include    "system.h"
 
-float tempCalc(int analogValue);
+int tempCalc(int analogValue);
 
 #endif
